refactor: rename log retention checker file and improve warning messages
